refactor: add Profile Header
