+++ src/styles/profile.css
@@ -1,0 +1,22 @@
+.profile-page {
+  position: relative;
+  width: 100%;
+  min-height: 100vh;
+  background-color: #221b18;
+  color: #c0c0c0;
+  padding: 40px 16px;
+}
+
+.profile-container {
+  max-width: 1024px;
+  margin: 0 auto;
+}
+
+.section-title {
+  font-size: 1.25rem; /* xl */
+  font-weight: 600;
+  color: #f5f5dc;
+  border-bottom: 2px solid #8b5a2b; /* RustBrown */
+  padding-bottom: 8px;
+  margin-bottom: 24px;
+}
